Make default IRule not implement IEnumerable<IRule>.
This was used to support cases like "2nd (monday OR friday) of the month, where just counting the 2nd occurrence of the rule would not be correct (e.g. first monday is the 1st occurrence, 1st friday is the 2nd occurrence, and therefore we report a match on the 1st friday)

Instead, have a separate interface `IHasAlternativeRules` only on types that can exhibit such behaviour, and where "the 2nd occurrence" should be counted for each of a set of alternate rules.
These types are presently: SetUnion and DaysOfWeek.